Remove gulp and ts-node dev dependencies
