--- gulp/GulpHelpers.ts
@@ -1,20 +1,0 @@
-/*
-* This program and the accompanying materials are made available under the terms of the
-* Eclipse Public License v2.0 which accompanies this distribution, and is available at
-* https://www.eclipse.org/legal/epl-v20.html
-*
-* SPDX-License-Identifier: EPL-2.0
-*
-* Copyright Contributors to the Zowe Project.
-*
-*/
-
-export interface ITaskFunction {
-    (done?: any): any;
-
-    description?: string;
-}
-
-export interface IGulpError extends Error {
-    showStack?: boolean;
-}
