+++ package.json
@@ -9,7 +9,7 @@
     "__tests__/__packages__/*"
   ],
   "scripts": {
-    "build": "lerna run build && gulp updateLicense && npm run lint && npm run checkTestsCompile && npm run circularDependencyCheck",
+    "build": "lerna run build && node scripts/updateLicenses.js && npm run lint && npm run checkTestsCompile && npm run circularDependencyCheck",
     "build:exe": "cd zowex && cargo build && cargo clippy && cargo test",
     "clean": "lerna run --parallel clean",
     "clean:exe": "cd zowex && cargo clean",
@@ -30,9 +30,7 @@
     "watch": "lerna run --parallel watch",
     "watch:exe": "cd zowex && cargo install cargo-watch && cargo watch -x build -x clippy -x test",
     "watch:test": "jest --watch",
-    "doc:clean": "rimraf docs/CLIReadme.md",
-    "doc:generate": "npm run doc:clean && gulp doc",
-    "generateCleanTypedoc": "npm run typedoc && gulp cleanTypedoc",
+    "generateCleanTypedoc": "npm run typedoc",
     "typedoc": "typedoc",
     "typedoc:packages": "lerna run --parallel typedoc",
     "audit:public": "npm audit --registry https://registry.npmjs.org/",
@@ -57,11 +55,6 @@
     "eslint-plugin-unused-imports": "^2.0.0",
     "fancy-log": "^1.3.3",
     "glob": "^7.1.6",
-    "gulp": "^4.0.2",
-    "gulp-cli": "^2.0.1",
-    "gulp-debug": "^4.0.0",
-    "gulp-plumber": "^1.2.1",
-    "gulp-replace": "^0.6.1",
     "husky": "^6.0.0",
     "jest": "^28.1.3",
     "jest-cli": "^29.0.0",
@@ -80,12 +73,8 @@
     "shebang-regex": "^2.0.0",
     "syncpack": "^5.6.10",
     "ts-jest": "^28.0.8",
-    "ts-node": "^7.0.1",
     "typedoc": "^0.23.10",
     "typescript": "^4.0.0"
-  },
-  "overrides": {
-    "glob-parent": "~5.1.2"
   },
   "jest": {
     "globals": {
